requirements.txt
# ==========================================
# Core Runtime
# ==========================================

numpy
torch
openai-whisper
ollama

# ==========================================
# Token Optimization
# ==========================================

tiktoken


